Tests/LibWeb/Text/expected/css/stroke-dasharray-animation.txt

same-length: 20px, 30px
lcm: 30px, 30px, 25px, 20px, 40px, 15px
none-list-low: none
none-list-high: 10px, 20px
list-none-high: none
